# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GroupFinder API",
    "version": "1.0.0",
    "description": "API documentation for the GroupFinder backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SendOtpRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          }
        }
      },
      "VerifyOtpRequest": {
        "type": "object",
        "required": [
          "email",
          "otp"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "SendOtpResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "OTP sent successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "expiresInSeconds": {
                "type": "integer",
                "example": 300
              }
            }
          }
        }
      },
      "VerifyOtpResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "OTP verified successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "example": "eyJhbGciOi..."
              },
              "tokenType": {
                "type": "string",
                "example": "Bearer"
              },
              "userId": {
                "type": "string",
                "example": "uuid-here"
              }
            }
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "Aarav"
          },
          "latitude": {
            "type": "number",
            "example": 12.9716
          },
          "longitude": {
            "type": "number",
            "example": 77.5946
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-here"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "Aarav"
          },
          "mobileNumber": {
            "type": "string",
            "nullable": true,
            "example": "9000000001"
          },
          "latitude": {
            "type": "number",
            "nullable": true,
            "example": 12.9716
          },
          "longitude": {
            "type": "number",
            "nullable": true,
            "example": 77.5946
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CurrentUserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Profile updated successfully"
          },
          "data": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
      "CreateEventRequest": {
        "type": "object",
        "required": [
          "category",
          "title",
          "description",
          "latitude",
          "longitude",
          "eventDate",
          "eventTime",
          "maxParticipants"
        ],
        "properties": {
          "category": {
            "type": "string",
            "example": "Tech"
          },
          "title": {
            "type": "string",
            "example": "React Builders Meetup"
          },
          "description": {
            "type": "string",
            "example": "A meetup for frontend enthusiasts."
          },
          "latitude": {
            "type": "number",
            "example": 12.9716
          },
          "longitude": {
            "type": "number",
            "example": 77.5946
          },
          "eventDate": {
            "type": "string",
            "example": "2026-04-10"
          },
          "eventTime": {
            "type": "string",
            "example": "10:00"
          },
          "maxParticipants": {
            "type": "integer",
            "example": 30
          }
        }
      },
      "CreateEventResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Event created successfully"
          },
          "data": {
            "$ref": "#/components/schemas/EventResponse"
          }
        }
      },
      "EventResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "creatorId": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "eventDate": {
            "type": "string",
            "format": "date"
          },
          "eventTime": {
            "type": "string"
          },
          "maxParticipants": {
            "type": "integer"
          },
          "currentParticipants": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SendJoinRequestResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Join request sent successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "eventId": {
                "type": "string"
              },
              "userId": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "example": "PENDING"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "ProcessRequestResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Request accepted successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "request": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              },
              "event": {
                "$ref": "#/components/schemas/EventResponse"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid OTP"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/send-otp": {
      "post": {
        "summary": "Send OTP",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendOtpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP and issue JWT",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOtpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "post": {
        "summary": "Create event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEventResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all events",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events/category/{category}": {
      "get": {
        "summary": "Get events by category",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events for a category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventByCategoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events/timing/{timing}": {
      "get": {
        "summary": "Get events by timing",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "timing",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "upcoming"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events for the selected timing bucket",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventByTimingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}/request": {
      "post": {
        "summary": "Send join request",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Join request sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendJoinRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/requests/{id}/accept": {
      "post": {
        "summary": "Accept request",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/requests/{id}/reject": {
      "post": {
        "summary": "Reject request",
        "tags": [
          "Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request rejected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}